A mostly-working Parser, with known gaps
There's a lot here.  The parser generates S-expressions for relatively
simply inputs.  It doesn't support "." (e.g. "(a . b)") nor does it
support quoted strings, but those can plausibly be said to be fringe
features.

Error-handling is AWFUL.  I'm actually not quite sure what sort of
information ought to be in error-reporting; there's no "error"
S-expression.  Should there be?  I'm not sure.

But the parser works, and that's a lot.