Extend/perfect evaluation
We can evaluate pretty complicated-looking strings of Schemelike
programs.  There are issues, to be sure, but we can see a way forward.
Writing new evaluators for new builtins is going to be boring and could
probably be abstracted.

Numbers turn out to be hard.  The string-representation and the fact
that `float`s and `int`s aren't even _comparable_ in Go means it's going
to be hard to do arithmetic.  For now, I'll probably say ℤ is a big
enough ring to do all the arithmetic we really want to do.