Simplify the parser's stack
Duh.  A stack-of-sexpr-lists is too much.  We just need a stack with a
marker so that "pop" doesn't go too far.  Since "Sexpr" is a generic
interface, we can introduce these false S-expressions into the parser's
internals.

We should be careful to not let them leak out, though.

The error messages for parse-errors was pretty non-existent.  It's still
"pretty" non-existent, but there's something very basic and a
promise/hope to do better.