Implement "quote", more or less correctly
The Parser gets some new state.  In addition to the stack of
S-expressions, there's a flag to say whether "quote" is active.  Those
two little bits seem reasonable for now.  If the parser gets too much
more, I might want to change the simple "run" method to allow
transitions.  But for now, the only states are "quote" and "not quote",
and the stack comes along as data.  Sort of.

This simplifies, greatly, the logic around ingesting symbols.  The
rules (hey, there are rules!  cf.

  https://www.scheme.com/tspl4/grammar.html#grammar:symbols

) are pretty premissive.  I only support a subset of that here, but now
there's "isPartOfASymbol" that we can make as complicated as we like.
For when people want to run esoteric production code in scam instead of
just installing chezscheme.

The lexer got some improvements.  QUOTE is just a token now, so that's
easy.  And, since "#t" and "#f" would be symbols except that they start
with the wrong initial thing, I just put them up higher in the switch
statement, so booleans work.  And looking for numbers is still "the
same", just using the helper functions to make things differently-clear.