The output of "scheme-01-toys".  Working
This file was generated by

  go run scam.go -in examples/scheme-01-toys.ss > examples/scheme-01-toys.out

We save it for posterity.  It works.  I'll check it for correctness,
maybe.

The rendering of Cons cells is irritating, but I think it needs to stay
this way for now.  At least until I fee a whole lot better about how the
evaluator works; making them pretty will mask issues.