Refactor evaluators, removing "sexpr_error"
The evaluators are a little simpler.  It's sort of obvious how to write
new ones, with "requireArgCount", then just treat the list as an array.
I'll start adding evaluation contexts and symbol tables in a bit.

As a happy side-effect, I figured out how to make "+" (and really any
arithmetic) work for floats, because ParseInt and ParseFloat both
return errors if they can't parse the string.  That's a sort-of
introspection we end up using to figure out what kind of accumulator to
use.