Correct parsing of '(a b) (and QUOTE in general)
The trick is to not apply the quote until we've got the correct
S-expression in hand.  Unlike rev. eb3a2ed, this code uses the stack
itself to mark when the context is correct.  The older version just used
a boolean to flag the state "there's a quote symbol somewhere down below
you", and that's wrong.  The thing we want to know is "there's a quote
symbol right here", so the stack is the right place.