A flawed-but-plausible implementation of "let"
I write this down for posterity (and in case I end up biffing the next
step).

This (non-)implementation shows that the evaluation context shouldn't
actually be the stack but just the head of the stack.  It's not a global
thing we pass around; it's just the immediate context that we want.
Otherwise, either changes leak or we're obliged to pop after evaluation.