Fix a bug in the definition of lambda
Applying a λ expression ended up evaluating its arguments twice.  That's
not what we wanted!  A function is given is arguments
pre-evaluated (that's the macro/function confusion of rev. 127b673, just
in reverse), so there's no need to call "evaluate".

All tests are green, and

  go run scam.go -in examples/scheme-02-recursion.ss

is very close to evaluating error-free.