Refactor primitives to be true functions
A function is a thing that is applied to already-evaluated arguments.  A
macro is something that gets to manipulate the evaluation context.  Some
of the primitives are macros, but most of them are functions.  Now
they're separate.

λ is a macro that evaluates to a function.  It's cool.  The others are
cool, too.  But λ is the coolest.

I don't think there's a macro that evaluates to a macro.  Maybe
"define-syntax", but I really doubt I'll get all the way there.