Implement "-" and deobfuscate arithmetic reduce
The "reduce" method had gotten more than a little out of hand.  It's a
lot simpler now.  If we treat "intOrFloat" as a number-type in its own
right, and define "+=" and "-=" (and "*=" and "/=" if we get there),
/and/ define helpers to convert back and forth to strings and Sexprs,
that's the place to stick the complicated bookkeeping stuff.

Now it's simpler.

There's an issue in "mkNaryFn", namely that it's not enforcing the
argument-count.  So I'll fix that, and we'll be good.