The output of "scheme-03-numbers.ss", working
I didn't do "-in" because the lines get messed up.  Instead, I typed in
each command and then just copied the whole thing.  Notice that all the
answer are correct.  I did a couple of mulit-step inputs, because my
Emacs-fu isn't sufficient to send S-expressions from one buffer to
another.  There's probably a way to do that, right.

It's lovely.