The applicative Y-combinator, in action
This is from the end of Chapter 9.  The Y-combinator managed to define a
"length" function out of thin air.  You want to define "length"?  Well,
you know the length of the empty list.  Just hypothesize the existence
of a function that will compute the length of any list shorter than the
one it has, and also knows how to compute the empty list, and call it.

Or some sort of magic like that.  It's spooky.

You can also define a two-argument "rember", but it needs to be Curried.
I'm not quite sure why.

But the point of this exercise isn't to learn about recursion,
computability, and combinators.  It's to marvel in the fact that SCAM
can define all these things and they work.