A final snapshot of "Chapter 9"
The file "scheme-09-again.out" is the transcript from when SCAM runs in
Emacs as an "inferior scheme".  It's not a particularly pejorative term;
SCAM is clearly a find Scheme that can even implement a Y-combinator.

I also included two sample documents that depict _me_ working through
the examples to re-derive the Y-combinator.